data-backend/requirements.txt
# Django and REST Framework
Django>=5.0.0
djangorestframework>=3.14.0
django-filter>=23.0
django-cors-headers>=4.0.0

# Authentication
djangorestframework-simplejwt>=5.3.0
django-allauth>=0.57.0
dj-rest-auth>=5.0.0
cryptography>=41.0.0

# Database drivers
# SQLite is built-in with Python

# Search engines
meilisearch>=0.28.0

# Graph database
neo4j>=5.0.0

# Vector database and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Token counting
tiktoken>=0.5.0

# Progress bars
tqdm>=4.65.0

# Utilities
python-dateutil>=2.8.0
markdown>=3.4.0

# Vector service
flask>=3.0.0
requests>=2.31.0
